# perf: remove unnecessary c++ call with url.format