--- a/lib/fetch/response.js
+++ b/lib/fetch/response.js
@@ -117,7 +117,7 @@ class Response {
     responseObject[kState].status = status
 
     // 6. Let value be parsedURL, serialized and isomorphic encoded.
-    const value = isomorphicEncode(URLSerializer(parsedURL))
+    const value = isomorphicEncode(parsedURL.toString())
 
     // 7. Append `Location`/value to responseObject’s response’s header list.
     responseObject[kState].headersList.append('location', value)